find correlated variables

This function takes a data frame or a matrix and returns either a data frame or a list of data frames with correlated variables

func_correlation(
  data,
  target = NULL,
  correlation_thresh = NULL,
  use_obs = NULL,
  correlation_method = NULL
)

Arguments

data
either a data frame or a matrix

target
either a string (name of the predictor/response in the data set) or a vector of strings (predictor/response names of the data set)

correlation_thresh
a float indicating the correlation threshold

use_obs
one of "everything", "all.obs", "complete.obs", "na.or.complete", "pairwise.complete.obs"

correlation_method
one of "pearson", "kendall", "spearman"

Value

either a data frame or a list of data frames

Details

This function takes a data frame or a matrix and returns the correlated variables of the data